+++ drivers/interrupts/keyboard.c
@@ -1,4 +1,4 @@
-#include "../drivers/io.h"
+#include "../io/io.h"
 
 #define KEYBOARD_DATA_PORT 0x60
