+++ drivers/interrupts/pic.c
@@ -1,4 +1,4 @@
-#include "../drivers/io.h"
+#include "../io/io.h"
 #include "pic.h"
 
 /*
