+++ kmain.c
@@ -1,6 +1,6 @@
-#include "drivers/serial_port.c"
-#include "segmentation/memory_segments.h"
-#include "interrupt/interrupts.h"
+#include "drivers/serial_port/serial.c"
+#include "memory/segmentation/memory_segments.h"
+#include "drivers/interrupts/interrupts.h"
 #include "multiboot.h"
 
 /* Function to initialize */
